requirements.txt
adbutils==0.11.0
Django==4.1.6
djangorestframework==3.14.0
elasticsearch==8.1.2
Jinja2==3.0.3
pandas==1.4.2
python-dotenv==1.0.0
PyYAML==6.0
requests==2.26.0
uiautomator2==2.16.6
